+++ etc/sync-function-loader.js
@@ -2,13 +2,21 @@ var fs = require('fs');
 var path = require('path');
 var indent = require('../lib/indent.js/indent.min.js');
 
-exports.load = function(docDefinitionFilename, baseDir) {
-  var syncFuncTemplateDir = baseDir ? baseDir + '/etc' : 'etc';
+/**
+ * Generates a complete sync function from the specified document definitions file.
+ *
+ * @param {string} docDefinitionsFile The path to the document definitions file
+ *
+ * @returns The full contents of the sync function generated from the specified document definitions file
+ */
+exports.load = function(docDefinitionsFile) {
+  var syncFuncTemplateDir = path.dirname(module.filename);
+
   function readSyncFunctionFragment(fullMatch, fragmentFilename) {
     return readFileFragment(fullMatch, fragmentFilename, syncFuncTemplateDir);
   }
 
-  var docDefinitionDir = path.dirname(docDefinitionFilename);
+  var docDefinitionDir = path.dirname(docDefinitionsFile);
   function readDocDefinitionFragment(fullMatch, fragmentFilename) {
     return readFileFragment(fullMatch, fragmentFilename, docDefinitionDir);
   }
@@ -29,7 +37,7 @@ exports.load = function(docDefinitionFilename, baseDir) {
 
   var syncDocDefn;
   try {
-    syncDocDefn = fs.readFileSync(docDefinitionFilename, 'utf8').trim();
+    syncDocDefn = fs.readFileSync(docDefinitionsFile, 'utf8').trim();
   } catch (ex) {
     if (ex.code === 'ENOENT') {
       console.log('ERROR: Sync document definitions file does not exist');