+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- [#123](https://github.com/Kashoo/synctos/issues/123): Specify sync-function-loader.js as the main package file
+
 ### Changed
 - [#118](https://github.com/Kashoo/synctos/issues/118): Embed indent.js as a static dependency